Fix Cluster.clusters
Implement visualisation of the results of this function.  When hovering
pointer over an entity lines are drawn to all clusters that will
influence it.

Implement Cluster.location that gives an absolute location of an entity
in the cluster. Returns Nothing if the entity is not in the cluster.

Remove unused Main.viewCursor function.