Implement animation
Hide cluster boundries - they were too noisy.

Generate an array of 20 x 10 entities on init.

The code to calculate forces is quite messy and probably can be
optimized, but seems to work correctly and performance is decent for 200
entities. Above that starts to deteriorate. As always, GC is consuming
the most of the time.