Re-implement binary particles example in Demo.elm
Compared to the previous (all in one) implementation in current master
branch the new one is significantly slower (in terms of performance). I
don't have any better numbers, but on my MacBook Pro 2015 the master
implementation runs smoothly (~30fps) with ~300 entities. The new
implementation chokes on anything above 100 entities.

This commit also removes old Surface and Main modules. I'm going to
iterate on Ecosystem and Demo so the code there was dead and rotting
fast.