Simplify ActorUpdate, Spawn and Init; put demos in own directory
The ActorUpdate and Spawn are no longer tagged union types. It was
necessary for ActorUpdate before because it was recursive. Now it's not
thnanks to separate Spawn type. So both of them are now simple
records.

Also init now takes a List (Spawn actor action). Previously it took
`List (ActorUpdate actor action)` which didn't really made sense. So
it's better now.

Finally there are two demos now: main demo and one I've implemented
during my FPAms talk. They have their own directory and are separately
compiled by the dist and develop targets of make. This will ensure that
demos are always up to date. In the future there probably should be a
separate demos target and also html and js code should have access to
the list of demos. But it's a start.