Introduce lazy evaluation of environment
There is a new module called Environment which exposes a type of the
same name. The updateActor function will now be called with a value of
type `Environment actor action` as the third and last argument (as desribed in ReadMe). This value allows actor to get information about:

1. Other actors around

   like previous groups argument. Now it's:

      Environment.actors environment

2. Particular actors

      Environment.actor id environment

3. Time since last update

      Environment.latency environment

4. Incoming interactions

      Environment.interactions environment

Conceptually it's a snapshot of the ecosystem state scoped for a
particular actor. Thanks to this design potentially heavy computation is
only performed on demand (e.g. a dead actor may not care about other
actors) and API gets simplified. The developer no longer needs the
injected inspect function or three other arguments passed to the
updateActor function. Everything is available on demand in this one
argument.

For technical reasons Interaction module also had to be created, as it
is shared between Ecosystem and Environment. I think it makes a nice API
this way.