Simplify demos loading by using URL query parameters
There is only one HTML file generated from index.pug. This file contains
links to itself with different value of "demo" query parameter (e.g.
"?demo=Binary"). The trick depends on the fact that changing query
parameter triggers full page refresh in a browser.

When the page is loaded with a parameter set, it's value is used to load
one of the demos (from src/Demos/*.elm).

Adding new demo will require to:

1. Add new link with a query parameter
2. Add a corresponding item to Demos enum in index.ts
3. Add a branch to switch statement in the load function in index.ts