Implement tracking feature, change key bindings
Clicking on an actor selects it. Selected actor is tracked - i.e. the
viewbox follows it, keeping it in the center.

Also, since space is bound to scroll down in most browsers, it is not a
good key for play / pause. So now "p" key is for play / pause, while
shift-p for stepping.